+++ CHANGELOG.md
@@ -6,6 +6,9 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 
 ## [unreleased]
 
+### Add
+- Add `reset filters` button in DwC import task [#2730]
+
 ### Changed
 - Add space on navbar in New taxon name [#2701]
 - Updated ruby gems and node packages.
@@ -45,6 +48,7 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 [#2716]: https://github.com/SpeciesFileGroup/taxonworks/issues/2716
 [#2721]: https://github.com/SpeciesFileGroup/taxonworks/pull/2721
 [#2724]: https://github.com/SpeciesFileGroup/taxonworks/pull/2724
+[#2730]: https://github.com/SpeciesFileGroup/taxonworks/issues/2730
 [#2731]: https://github.com/SpeciesFileGroup/taxonworks/issues/2731
 
 ## [0.22.3] - 2021-12-03